Code revision changes - merged from feat/nerf_tmp
