--- kornia/geometry/nerf/types.py
@@ -1,5 +1,0 @@
-from typing import Union
-
-from torch import device
-
-Device = Union[str, device]
