+++ test/nerf/cameras.txt
@@ -1,0 +1,4 @@
+# Camera list with one line of data per camera:
+#   CAMERA_ID, MODEL, WIDTH, HEIGHT, PARAMS[]
+# Number of cameras: 1
+1 SIMPLE_PINHOLE 1008 756 845.56535484459982 504 378
